export handler to mod.ts
