+++ .github/workflows/cargo.yml
@@ -12,19 +12,3 @@ jobs:
           toolchain: stable
       - run: cargo fmt --all --check
 
-  check:
-    runs-on: ubuntu-latest
-    steps:
-      - uses: actions/checkout@v2
-      - uses: actions/cache@v2
-        with:
-          path: |
-            ~/.cargo/registry
-            ~/.cargo/git
-            target
-          key: ${{ runner.os }}-cargo-check
-          restore-keys: ${{ runner.os }}-cargo
-      - uses: actions-rs/toolchain@v1
-        with:
-          toolchain: stable
-      - run: cargo check
